# optimize root hash calculation with adaptive parallelization