--- a/README.md
+++ b/README.md
@@ -150,6 +150,12 @@ The service is configured via environment variables:
 |----------|-------------|---------|
 | `BATCH_LIMIT` | Maximum number of commitments to process per batch | `1000` |
 | `ROUND_DURATION` | Duration between block creation rounds | `1s` |
+| `SMT_MAX_GOROUTINES` | Maximum concurrent goroutines for SMT hash calculations | CPU-based (2×cores, min 8, max 512) |
+
+**SMT Goroutine Configuration:**
+- **Not set** or **negative values**: Uses CPU-based defaults (2× CPU cores, minimum 8, maximum 512)
+- **0**: Sequential processing (no parallel goroutines)  
+- **Positive values**: Uses specified number of goroutines for parallel processing
 
 ## API Endpoints