Root hash calcualtion performance was evaluated by utilizing BenchmarkSMTRootHashOnly test. Additionally, hash caching was disabled in code to have full tree hash calcualtion on every test run.

Root hash calculation performance with sequential approach:
```
goos: linux
goarch: amd64
pkg: github.com/unicitynetwork/aggregator-go/internal/smt
cpu: 12th Gen Intel(R) Core(TM) i7-1255U
BenchmarkSMTRootHashOnly/HashOnly_1k-12                     1494            729284 ns/op          614990 B/op      14994 allocs/op
BenchmarkSMTRootHashOnly/HashOnly_5k-12                      298           3896364 ns/op         3071852 B/op      74996 allocs/op
BenchmarkSMTRootHashOnly/HashOnly_10k-12                     152           7791941 ns/op         6151860 B/op     149996 allocs/op
BenchmarkSMTRootHashOnly/HashOnly_100k-12                     12          86507463 ns/op        61591853 B/op    1499996 allocs/op
PASS
ok      github.com/unicitynetwork/aggregator-go/internal/smt    8.171s
```

Root hash calculation performance using parallel approach:
```
goos: linux
goarch: amd64
pkg: github.com/unicitynetwork/aggregator-go/internal/smt
cpu: 12th Gen Intel(R) Core(TM) i7-1255U
BenchmarkSMTRootHashOnly/HashOnly_1k-12                     2454            463138 ns/op          687334 B/op      17996 allocs/op
BenchmarkSMTRootHashOnly/HashOnly_5k-12                      489           2377031 ns/op         3434489 B/op      90021 allocs/op
BenchmarkSMTRootHashOnly/HashOnly_10k-12                     250           4785997 ns/op         6879210 B/op     180070 allocs/op
BenchmarkSMTRootHashOnly/HashOnly_100k-12                     19          69916767 ns/op        68950190 B/op    1801643 allocs/op
PASS
ok      github.com/unicitynetwork/aggregator-go/internal/smt    7.826s
```